+++ website/src/components/SearchPage/SearchFullUI.tsx
@@ -1,45 +1,32 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 
 import { DownloadDialog } from './DownloadDialog/DownloadDialog.tsx';
 import { DownloadUrlGenerator } from './DownloadDialog/DownloadUrlGenerator.ts';
 import { LinkOutMenu } from './DownloadDialog/LinkOutMenu.tsx';
 import { FieldFilterSet, SequenceEntrySelection, type SequenceFilter } from './DownloadDialog/SequenceFilters.tsx';
 import { RecentSequencesBanner } from './RecentSequencesBanner.tsx';
+import { SearchErrorDisplay } from './SearchErrorDisplay';
 import { SearchForm } from './SearchForm';
 import { SearchPagination } from './SearchPagination';
 import { SeqPreviewModal } from './SeqPreviewModal';
 import { Table, type TableSequenceData } from './Table';
+import { buildSequenceCountText } from './searchHelpers';
 import { stillRequiresSuborganismSelection } from './stillRequiresSuborganismSelection.tsx';
-import useQueryAsState, { type QueryState } from './useQueryAsState';
+import type { QueryState } from './useQueryAsState';
+import { useSearchLapisQueries } from './useSearchLapisQueries';
+import { useSearchState } from './useSearchState';
 import { getLapisUrl } from '../../config.ts';
-import useUrlParamState from '../../hooks/useUrlParamState';
-import { lapisClientHooks } from '../../services/serviceHooks.ts';
 import { DATA_USE_TERMS_FIELD, pageSize } from '../../settings';
 import type { Group } from '../../types/backend.ts';
 import type { LinkOut } from '../../types/config.ts';
-import {
-    type FieldValues,
-    type FieldValueUpdate,
-    type Schema,
-    type SequenceFlaggingConfig,
-    type SetSomeFieldValues,
-} from '../../types/config.ts';
-import { type OrderBy, type OrderDirection } from '../../types/lapis.ts';
+import { type FieldValues, type Schema, type SequenceFlaggingConfig } from '../../types/config.ts';
+import { type OrderBy } from '../../types/lapis.ts';
 import type { ReferenceGenomesLightweightSchema } from '../../types/referencesGenomes.ts';
 import type { ClientConfig } from '../../types/runtimeConfig.ts';
-import { formatNumberWithDefaultLocale } from '../../utils/formatNumber.tsx';
-import {
-    COLUMN_VISIBILITY_PREFIX,
-    getColumnVisibilitiesFromQuery,
-    getFieldVisibilitiesFromQuery,
-    MetadataFilterSchema,
-    NULL_QUERY_VALUE,
-    VISIBILITY_PREFIX,
-} from '../../utils/search.ts';
+import { MetadataFilterSchema } from '../../utils/search.ts';
 import { EditDataUseTermsModal } from '../DataUseTerms/EditDataUseTermsModal.tsx';
 import { ActiveFilters } from '../common/ActiveFilters.tsx';
-import ErrorBox from '../common/ErrorBox.tsx';
 import { type FieldItem, FieldSelectorModal } from '../common/FieldSelectorModal.tsx';
 
 export interface InnerSearchFullUIProps {
@@ -59,16 +46,6 @@ export interface InnerSearchFullUIProps {
     linkOuts?: LinkOut[];
 }
 
-const buildSequenceCountText = (totalSequences: number | undefined, oldCount: number | null, initialCount: number) => {
-    const sequenceCount = totalSequences ?? oldCount ?? initialCount;
-
-    const formattedCount = formatNumberWithDefaultLocale(sequenceCount);
-    const pluralSuffix = sequenceCount === 1 ? '' : 's';
-
-    return `Search returned ${formattedCount} sequence${pluralSuffix}`;
-};
-
-/* eslint-disable @typescript-eslint/no-unsafe-member-access -- TODO(#3451) this component is a mess a needs to be refactored */
 export const InnerSearchFullUI = ({
     accessToken,
     referenceGenomeLightweightSchema,
@@ -87,9 +64,7 @@ export const InnerSearchFullUI = ({
 }: InnerSearchFullUIProps) => {
     hiddenFieldValues ??= {};
 
-    const metadataSchema = schema.metadata;
-    const filterSchema = useMemo(() => new MetadataFilterSchema(metadataSchema), [metadataSchema]);
-
+    const filterSchema = useMemo(() => new MetadataFilterSchema(schema.metadata), [schema.metadata]);
     const [isColumnModalOpen, setIsColumnModalOpen] = useState(false);
 
     const columnFieldItems: FieldItem[] = useMemo(
@@ -106,188 +81,28 @@ export const InnerSearchFullUI = ({
         [schema.metadata, schema.primaryKey],
     );
 
-    const [state, setState] = useQueryAsState(initialQueryDict);
-
-    const [previewedSeqId, setPreviewedSeqId] = useUrlParamState<string | null>(
-        'selectedSeq',
-        state,
-        null,
-        setState,
-        'nullable-string',
-        (value) => !value,
-    );
-    const [previewHalfScreen, setPreviewHalfScreen] = useUrlParamState(
-        'halfScreen',
-        state,
-        false,
-        setState,
-        'boolean',
-        (value) => !value,
-    );
-    const [selectedSuborganism, setSelectedSuborganism] = useUrlParamState<string | null>(
-        schema.suborganismIdentifierField ?? '',
-        state,
-        null,
-        setState,
-        'nullable-string',
-        (value) => value === null,
-    );
-
-    const searchVisibilities = useMemo(() => {
-        return getFieldVisibilitiesFromQuery(schema, state);
-    }, [schema, state]);
-
-    const columnVisibilities = useMemo(() => getColumnVisibilitiesFromQuery(schema, state), [schema, state]);
-
-    const columnsToShow = useMemo(() => {
-        return schema.metadata
-            .filter((field) => columnVisibilities.get(field.name) === true)
-            .map((field) => field.name);
-    }, [schema.metadata, columnVisibilities]);
-
-    let orderByField = state.orderBy ?? schema.defaultOrderBy;
-    if (!columnsToShow.includes(orderByField)) {
-        orderByField = schema.primaryKey;
-    }
-
-    const orderDirection = state.order ?? schema.defaultOrder;
-
-    const page = parseInt(state.page ?? '1', 10);
-
-    const setPage = useCallback(
-        (newPage: number) => {
-            setState((prev: QueryState) => {
-                if (newPage === 1) {
-                    const withoutPageSet = { ...prev };
-                    delete withoutPageSet.page;
-                    return withoutPageSet;
-                } else {
-                    return {
-                        ...prev,
-                        page: newPage.toString(),
-                    };
-                }
-            });
-        },
-        [setState],
-    );
-
-    const setOrderByField = (field: string) => {
-        setState((prev: QueryState) => ({
-            ...prev,
-            orderBy: field,
-            page: '1',
-        }));
-    };
-
-    const setOrderDirection = (direction: OrderDirection) => {
-        setState((prev: QueryState) => ({
-            ...prev,
-            order: direction,
-            page: '1',
-        }));
-    };
-
-    /**
-     * The `fieldValues` are the values of the search fields.
-     * The values are initially loaded from the default values set in `hiddenFieldValues`
-     * and the initial `state` (URL search params).
-     */
-    const fieldValues = useMemo(() => {
-        return filterSchema.getFieldValuesFromQuery(state, hiddenFieldValues);
-    }, [state, hiddenFieldValues, filterSchema]);
-
-    /**
-     * Update field values (query parameters).
-     * If value is '' or null, the query parameter is unset.
-     * For multi-select fields, we handle fieldValuesToSet as an array where:
-     * - If value is an array, it sets multiple values for that field
-     * - If value is '' or null, it clears the field
-     */
-    const setSomeFieldValues: SetSomeFieldValues = useCallback(
-        (...fieldValuesToSet: FieldValueUpdate[]) => {
-            setState((prev: QueryState) => {
-                const newState = { ...prev };
-                fieldValuesToSet.forEach(([key, value]) => {
-                    if (value === '' || value === null) {
-                        if (Object.keys(hiddenFieldValues).includes(key)) {
-                            // keep explicitly empty fields because they override the hiddenFieldValues here
-                            newState[key] = '';
-                        } else {
-                            // we can delete keys that are not in the hiddenFieldValues
-                            delete newState[key];
-                        }
-                    } else if (Array.isArray(value)) {
-                        // Handle array values for multi-select
-                        if (value.length === 0) {
-                            delete newState[key];
-                        } else {
-                            newState[key] = value.map((v) => v ?? NULL_QUERY_VALUE);
-                        }
-                    } else {
-                        newState[key] = value;
-                    }
-                });
-                return newState;
-            });
-            setPage(1);
-        },
-        [setState, setPage, hiddenFieldValues],
-    );
-
-    const removeFilter = (metadataFilterName: string) => {
-        if (Object.keys(hiddenFieldValues).includes(metadataFilterName)) {
-            const hiddenValue = hiddenFieldValues[metadataFilterName];
-            // If it's an array with nulls, filter them out (shouldn't happen but TypeScript doesn't know)
-            const valueToSet = Array.isArray(hiddenValue)
-                ? hiddenValue.filter((v): v is string => v !== null)
-                : hiddenValue;
-            setSomeFieldValues([metadataFilterName, valueToSet]);
-        } else {
-            setSomeFieldValues([metadataFilterName, null]);
-        }
-    };
-
-    const setASearchVisibility = (fieldName: string, visible: boolean) => {
-        setState((prev: QueryState) => {
-            const newState = { ...prev };
-            const key = `${VISIBILITY_PREFIX}${fieldName}`;
-            const metadataField = schema.metadata.find((field) => {
-                let name = field.name;
-                if (field.rangeOverlapSearch) {
-                    name = field.rangeOverlapSearch.rangeName;
-                }
-                return name === fieldName;
-            });
-            const defaultVisible = metadataField?.initiallyVisible === true;
-            if (visible === defaultVisible) {
-                delete newState[key];
-            } else {
-                newState[key] = visible ? 'true' : 'false';
-            }
-            if (!visible) {
-                delete newState[fieldName];
-            }
-            return newState;
-        });
-        if (!visible) {
-            setPage(1);
-        }
-    };
-
-    const setAColumnVisibility = (fieldName: string, visible: boolean) => {
-        setState((prev: QueryState) => {
-            const newState = { ...prev };
-            const key = `${COLUMN_VISIBILITY_PREFIX}${fieldName}`;
-            const defaultVisible = schema.tableColumns.includes(fieldName);
-            if (visible === defaultVisible) {
-                delete newState[key];
-            } else {
-                newState[key] = visible ? 'true' : 'false';
-            }
-            return newState;
-        });
-    };
+    const {
+        previewedSeqId,
+        setPreviewedSeqId,
+        previewHalfScreen,
+        setPreviewHalfScreen,
+        selectedSuborganism,
+        setSelectedSuborganism,
+        searchVisibilities,
+        columnVisibilities,
+        columnsToShow,
+        orderByField,
+        orderDirection,
+        page,
+        setPage,
+        setOrderByField,
+        setOrderDirection,
+        fieldValues,
+        setSomeFieldValues,
+        removeFilter,
+        setASearchVisibility,
+        setAColumnVisibility,
+    } = useSearchState(initialQueryDict, schema, hiddenFieldValues);
 
     useEffect(() => {
         if (showEditDataUseTermsControls && dataUseTermsEnabled) {
@@ -302,10 +117,6 @@ export const InnerSearchFullUI = ({
         dataUseTermsEnabled,
         schema.richFastaHeaderFields,
     );
-
-    const hooks = lapisClientHooks(lapisUrl).zodiosHooks;
-    const aggregatedHook = hooks.useAggregated({}, {});
-    const detailsHook = hooks.useDetails({}, {});
 
     const [selectedSeqs, setSelectedSeqs] = useState<Set<string>>(new Set());
     const sequencesSelected = selectedSeqs.size > 0;
@@ -324,48 +135,25 @@ export const InnerSearchFullUI = ({
 
     const downloadFilter: SequenceFilter = sequencesSelected ? new SequenceEntrySelection(selectedSeqs) : tableFilter;
 
-    useEffect(() => {
-        aggregatedHook.mutate({
-            ...lapisSearchParameters,
-            fields: [],
-        });
-        const OrderByList: OrderBy[] = [
-            {
-                field: orderByField,
-                type: orderDirection,
-            },
-        ];
-        // @ts-expect-error because the hooks don't accept OrderBy
-        detailsHook.mutate({
-            ...lapisSearchParameters,
-            fields: [...columnsToShow, schema.primaryKey],
-            limit: pageSize,
-            offset: (page - 1) * pageSize,
-            orderBy: OrderByList,
-        });
-    }, [lapisSearchParameters, schema.tableColumns, schema.primaryKey, pageSize, page, orderByField, orderDirection]);
+    const {
+        aggregatedHook,
+        detailsHook,
+        totalSequences,
+        oldData,
+        oldCount,
+        firstClientSideLoadOfDataCompleted,
+        firstClientSideLoadOfCountCompleted,
+    } = useSearchLapisQueries(
+        lapisUrl,
+        lapisSearchParameters,
+        columnsToShow,
+        schema.primaryKey,
+        page,
+        orderByField,
+        orderDirection,
+    );
 
-    const totalSequences = aggregatedHook.data?.data[0].count ?? undefined;
     const linkOutSequenceCount = downloadFilter.sequenceCount() ?? totalSequences;
-
-    const [oldData, setOldData] = useState<TableSequenceData[] | null>(null);
-    const [oldCount, setOldCount] = useState<number | null>(null);
-    const [firstClientSideLoadOfDataCompleted, setFirstClientSideLoadOfDataCompleted] = useState(false);
-    const [firstClientSideLoadOfCountCompleted, setFirstClientSideLoadOfCountCompleted] = useState(false);
-
-    useEffect(() => {
-        if (detailsHook.data?.data && oldData !== detailsHook.data.data) {
-            setOldData(detailsHook.data.data);
-            setFirstClientSideLoadOfDataCompleted(true);
-        }
-    }, [detailsHook.data?.data, oldData]);
-
-    useEffect(() => {
-        if (aggregatedHook.data?.data && oldCount !== aggregatedHook.data.data[0].count) {
-            setOldCount(aggregatedHook.data.data[0].count);
-            setFirstClientSideLoadOfCountCompleted(true);
-        }
-    }, [aggregatedHook.data?.data, oldCount]);
 
     const showMutationSearch =
         schema.submissionDataTypes.consensusSequences &&
@@ -424,35 +212,7 @@ export const InnerSearchFullUI = ({
             >
                 <RecentSequencesBanner organism={organism} />
 
-                {(detailsHook.isError || aggregatedHook.isError) &&
-                    // @ts-expect-error because response is not expected on error, but does exist
-                    (aggregatedHook.error?.response?.status === 503 ? (
-                        <div className='p-3 rounded-lg text-lg text-gray-700 text-italic'>
-                            {' '}
-                            The retrieval database is currently initializing – please check back later.
-                        </div>
-                    ) : (
-                        <div className='bg-red-400 p-3 rounded-lg'>
-                            <p>There was an error loading the data</p>
-                            <details>
-                                <summary className='text-xs cursor-pointer py-2'>More details</summary>
-                                <p className='text-xs'>{JSON.stringify(detailsHook.error)}</p>
-
-                                <p>{detailsHook.error?.message}</p>
-                                <p>{aggregatedHook.error?.message}</p>
-                            </details>
-                        </div>
-                    ))}
-                {(detailsHook.isPaused || aggregatedHook.isPaused) &&
-                    (!detailsHook.isSuccess || !aggregatedHook.isSuccess) && (
-                        <ErrorBox title='Connection problem'>
-                            The browser thinks you are offline. This will affect site usage, and many features may not
-                            work. If you are actually online, please try using a different browser. If the problem
-                            persists, feel free to create an issue in{' '}
-                            <a href='https://github.com/pathoplexus/pathoplexus/issues'>our Github repo</a> or email us
-                            at <a href='mailto:bug@pathoplexus.org'>bug@pathoplexus.org</a>.
-                        </ErrorBox>
-                    )}
+                <SearchErrorDisplay detailsHook={detailsHook} aggregatedHook={aggregatedHook} />
 
                 <div
                     className={`
